// frontend/packages/types/api/models/Route.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Stop } from './Stop';
export type Route = {
    id: number;
    name: string;
    description: string;
    stops: Array<Stop>;
    createdAt: string;
    updatedAt: string;
};