// frontend/packages/types/api/models/Train.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Route } from './Route';
export type Train = {
    id: number;
    name: string;
    route: Route;
    scheduleDays: number;
    createdAt: string;
    updatedAt: string;
};